requirements.txt
openai>=2.15.0
python-dotenv==1.0.0
pytest>=7.0.0
tavily-python>=0.3.0
chromadb>=0.4.0
